requirements.txt
customtkinter>=5.2.0
requests>=2.28.0
clipboard>=0.0.4
py7zr>=0.20.0
rarfile>=4.0
Pillow>=10.0.0
